requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
uvicorn==0.17.6
alembic==1.7.7
fastapi==0.77.1
celery[redis]==5.2.7
python-dotenv==0.20.0